+++ spec/components/schemas/batches/base.ts
@@ -23,12 +23,6 @@ const base: SchemaObject = {
       description: 'Message channel',
       type: 'string',
     },
-    contacts: {
-      type: 'object',
-    },
-    columnMap: {
-      type: 'object',
-    },
   },
   required: [
     'name',
