+++ spec/components/schemas/batches/multipart-batch.ts
@@ -1,40 +1,24 @@
 // tslint:disable:max-line-length
 import { SchemaObject } from 'openapi3-ts';
 import { createComponentRef } from '../../../../utils/ref';
-import { ref as smsBatchRef } from './batch.sms';
-import { ref as whatsAppBatchRef } from './batch.whatsapp';
+import { ref as multipartSmsBatchRef } from './multipart/batch.sms';
+import { ref as multipartWhatsAppBatchRef } from './multipart/batch.whatsapp';
 
 const base: SchemaObject = {
   title: 'Batch Object',
   description: 'This is a Batch object model.',
-  allOf: [{
-    oneOf: [{
-      $ref: smsBatchRef,
-    }, {
-      $ref: whatsAppBatchRef,
-    }],
-    discriminator: {
-      propertyName: 'channel',
-      mapping: {
-        SMS: smsBatchRef,
-        WHATSAPP: whatsAppBatchRef,
-      },
-    },
+  oneOf: [{
+    $ref: multipartSmsBatchRef,
   }, {
-    type: 'object',
-    properties: {
-      contacts: {
-        type: 'object',
-        readOnly: true,
-      },
-      columnMap: {
-        type: 'object',
-      },
-    },
-    required: [
-      'columnMap',
-    ],
+    $ref: multipartWhatsAppBatchRef,
   }],
+  discriminator: {
+    propertyName: 'channel',
+    mapping: {
+      SMS: multipartSmsBatchRef,
+      WHATSAPP: multipartWhatsAppBatchRef,
+    },
+  },
 };
 
 export const ref = createComponentRef(__filename);
