+++ spec/components/schemas/batches/multipart/batch.whatsapp.ts
@@ -1,30 +1,25 @@
 // tslint:disable:max-line-length
 import { SchemaObject } from 'openapi3-ts';
-import { createComponentRef } from '../../../../utils/ref';
-import { ref as baseRef } from './base';
-import { ref as templateContentRef } from '../content/template';
+import { createComponentRef } from '../../../../../utils/ref';
+import { ref as multipartBaseRef } from './base';
+import { ref as templateContentRef } from './content/template';
 
 const obj: SchemaObject = {
   title: 'Batch Object',
   description: 'This is a Batch object model.',
   type: 'object',
   allOf: [{
-    $ref: baseRef,
+    $ref: multipartBaseRef,
   }, {
     type: 'object',
     properties: {
       message: {
         type: 'object',
         properties: {
-          from: {
-            title: 'Sender ID',
-            description: 'This is the identifier of sender of this message. The sender shoud be created with a credential.',
-            type: 'string',
-          },
           contents: {
             type: 'array',
             items: {
-              oneOf: [{
+              anyOf: [{
                 $ref: templateContentRef,
               }],
               discriminator: {
@@ -38,9 +33,6 @@ const obj: SchemaObject = {
         },
       },
     },
-    required: [
-      'message',
-    ],
   }],
 };
